requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.10.1
branca==0.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Django==4.2.6
folium==0.15.0
fonttools==4.44.0
google-analytics-data==0.17.2
google-api-core==2.12.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
guindex==0.0.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
seaborn==0.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
